Add two run configurations
The start config will just start the app locally. The build run config probably won't be useful all that often but it's what gets run when changes are pushed to GitHub. Should it fail, the build will fail. To save time, you can run it yourself to see if everything is ready to be pushed.