OH MY GOD I JUST WASTED 3 HOURS FIGURING THIS OUT
Apparently, the place where you call makeAutoObservable in the constructor is really important. I was calling it at the beginning but since the fields were technically uninitialized at that point, the values were never being observed properly and re-renders weren't happening even when the state changed (such as when a message was edited).