Fix channels not loading when switching Towers
As it turns out, the `useEffectOnce` was truly only being called once! Now we have a `useEffect` which depends on the Tower that is passed in. Should that Tower change, the hook will be called again should the Tower that is being viewed change.

Also, the `refreshChannels` flow has been renamed to `initializeChannels` to better reflect its purpose. It will only run once for a given Tower so that we aren't needlessly sending a request every time the user switches Towers.