Bugfix 1
Fix pfp not updating on the /me page but updating in the TowerList. The reason the picture would only update in the TowerList is because it was an observer() component, so it would get the updated URL provided by getAvatarUrl. However, the UserDetails component wouldn't because it wasn't wrapped with observer().